+++ src/breadcrumbs-header/styles.js
@@ -6,11 +6,12 @@ import TYPE_SCALE from '../constants/type-scale';
 import BORDER_WIDTH from '../constants/border-width';
 import COLORS from '../constants/colors';
 import FONT_WEIGHT from '../constants/font-weight';
+import MEDIA from '../constants/media-queries';
 
 const { GREY_100, GREY_10, WHITE } = COLORS;
 const { TS_2, TS_5 } = TYPE_SCALE;
 const { BW_1 } = BORDER_WIDTH;
-const { FW_700, FW_400 } = FONT_WEIGHT;
+const { FW_100, FW_400 } = FONT_WEIGHT;
 
 export const Header = styled.div`
   align-items: center;
@@ -21,7 +22,7 @@ export const Header = styled.div`
 `;
 
 export const StyledLink = styled(Link)`
-  ${FW_700};
+  ${FW_100};
   color: ${GREY_100};
   text-decoration: none;
 `;
@@ -33,6 +34,10 @@ export const Separator = styled.span`
 
 export const Headline = styled.h1`
   ${FW_400};
-  ${TS_2};
+  ${TS_5};
   color: ${GREY_100};
+
+  ${MEDIA.small`
+    ${TS_2};
+  `};
 `;